refactor & decouple & fix
